Add files via upload
The methodArifinMod24.m file contains the Matlab code of the modified A.Z. Arifin and A. Asosno method of agglomerative multilevel thresholding of grayscale images. The modification consists in follows. Firstly, the distance between pairs of adjacent clusters is calculated using the formula for the increment of the total squared error. Secondly, the quality of each partition into clusters is assessed. 
The program receives an image at the input and checks whether it is 8-bit or 24-bit. In any case, it converts it to grayscale. 
During the execution of the main loop, the values are calculated and the partitions are generated. 
At the end, a graph of the approximation error values is displayed in linear and logarithmic scales. 
The program main loop is linear. One partition is generated per iteration. Working with large images, there is a slowdown due to the fact that the standard data output function in the Matlab environment takes a significant amount of time (it seems that it is not linear). It should be recoded if possible.
A way out of this situation is to output only a set of individual partitions, for example, from 15 to 1 cluster. Practice shows that the human visual perception experiences difficulties in distinguishing more than 15 different clusters (shades of gray) in one partition. 

10/31/2024 			Khanykov I.G.


Файл methodArifinMod24.m содержит Matlab код модифицированного метода Арифина. Модификация заключается в том, что, во-первых, расстояние между парами смежных кластеров рассчитывается по формуле приращения суммарной квадратичной ошибки. Во-вторых, оценивается качество каждого разбиения на кластеры. 
На входе программа принимает изображение и проверят его, является ли оно 8-битным или 24-битным. В любом случае конвертирует в оттенки серого.
В ходе исполнения основного цикла происходит расчет значений и генерация разбиений. 
В конце выводится график значений ошибок аппроксимаций в линейном и логарифмическом масштабе. Цикл программы линейный. За одну итерацию генерируется одно разбиение. 
При работе с большими изображениями наблюдается торможение из-за того, что стандартная функция вывода данных в среде Matlab затрачивает существенное время (похоже на то, что она не линейная). Ее нужно по возможности переписать. 
Вариант выхода из сложившейся ситуации – выводить только набор отдельных разбиений, например, от 15 до 1 кластера. Практика показывает, что больше 15 различных кластеров (оттенков серого) в одном разбиении человеческий глаз трудно улавливает. 

31.10.2024 				Ханыков И.Г.